Create base class for client testing
