+++ http-api/src/test/java/net/runelite/http/api/AbstractApiClientTest.java
@@ -1,0 +1,94 @@
+/*
+ * Copyright (c) 2020, Ugnius <https://github.com/UgiR>
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without
+ * modification, are permitted provided that the following conditions are met:
+ *
+ * 1. Redistributions of source code must retain the above copyright notice, this
+ *    list of conditions and the following disclaimer.
+ * 2. Redistributions in binary form must reproduce the above copyright notice,
+ *    this list of conditions and the following disclaimer in the documentation
+ *    and/or other materials provided with the distribution.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ * ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+ * WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+ * DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+ * (INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+ * LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ * ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+ * (INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+ * SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
+ */
+package net.runelite.http.api;
+
+import java.io.IOException;
+import okhttp3.Interceptor;
+import okhttp3.MediaType;
+import okhttp3.OkHttpClient;
+import okhttp3.Protocol;
+import okhttp3.Request;
+import okhttp3.Response;
+import okhttp3.ResponseBody;
+import okhttp3.mockwebserver.MockWebServer;
+import org.junit.After;
+import org.junit.Before;
+import org.junit.Rule;
+import org.junit.rules.ExpectedException;
+
+public abstract class AbstractApiClientTest
+{
+	public final MockWebServer server = new MockWebServer();
+	protected OkHttpClient.Builder testClient;
+
+	@Rule
+	public final ExpectedException exception = ExpectedException.none();
+
+	@Before
+	public void before() throws IOException
+	{
+		server.start();
+		testClient = new OkHttpClient.Builder()
+			.addInterceptor(chain ->
+			{
+				Request request = chain.request()
+					.newBuilder()
+					.url(server.url("/"))
+					.build();
+
+				return chain.proceed(request);
+			});
+	}
+
+	@After
+	public void after() throws IOException
+	{
+		server.shutdown();
+	}
+
+	// Captures the request at this interceptor stage and short circuits the response by returning a dummy response
+	public static class CaptureRequestInterceptor implements Interceptor
+	{
+
+		private Request request;
+
+		@Override
+		public Response intercept(Chain chain) throws IOException
+		{
+			this.request = chain.request().newBuilder().build();
+			return new Response.Builder()
+				.body(ResponseBody.create(MediaType.parse("application/json"), ""))
+				.code(200)
+				.protocol(Protocol.HTTP_2)
+				.request(chain.request())
+				.build();
+		}
+
+		public Request getRequest()
+		{
+			return request;
+		}
+	}
+}
